# openapi.yaml
openapi: 3.2.0
info:
  title: GREENHERB API
  description: Plataforma Web de Gestão Inteligente de uma Estufa
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /lotes:
    get:
      tags:
        - Lotes de Cultivo
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
    post:
      tags:
        - Lotes de Cultivo
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoteCultivo"
      responses:
        "201":
          description: Created
  /lotes/exportar:
    get:
      tags:
        - Lotes de Cultivo
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /lotes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags:
        - Lotes de Cultivo
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoteCultivo"
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
...
  /medicoes:
    get:
      tags:
        - Monitorização
      security:
        - cookieAuth: []
      parameters:
        - name: loteId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags:
        - Monitorização
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicaoAmbiental"
      responses:
        "201":
          description: Created
  /medicoes/exportar:
    get:
      tags:
        - Monitorização
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /medicoes/alertas:
    get:
      tags:
        - Monitorização
      security:
        - cookieAuth: []
      responses:
        "200":
          description: OK
  /medicoes/alertas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    patch:
      tags:
        - Monitorização
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                justificacao:
                  type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access
  schemas:
...
    LoteCultivo:
      type: object
      required: [ervaId, planoId, quantidadeInicial]
      properties:
        ervaId:
          type: string
        planoId:
          type: string
        estado:
          type: string
          enum: [Ativo, Concluído, Comprometido]
        modo:
          type: string
          enum: [Manual, Automático]
        dataInicio:
          type: string
          format: date-time
        dataFimReal:
          type: string
          format: date-time
        quantidadeInicial:
          type: number
        perdas:
          type: number
        loteOrigemId:
          type: string
...
    MedicaoAmbiental:
      type: object
      required: [loteId]
      properties:
        loteId:
          type: string
        temperatura:
          type: number
        humidade:
          type: number
        luminosidade:
          type: number
        dataHora:
          type: string
          format: date-time
...